use service oriented architecture for authentication
