+++ src/services/interfaces/iauthentication.service.ts
@@ -1,0 +1,7 @@
+import { Request, Response } from 'express';
+
+export interface IAuthenticationService {
+    register(req: Request, res: Response): Promise<Response>;
+    login(req: Request, res: Response): Promise<Response>;
+    logout(req: Request, res: Response): Promise<Response>;
+}
